readds qr code logic and moves functions to relevant files
