--- src/routes/userRoutes.ts
@@ -1,8 +1,0 @@
-import { Router } from 'express';
-import { getUsers } from '../controllers/userController';
-
-const router = Router();
-
-router.get('/users', getUsers);
-
-export default router;
