+++ src/types/express-session.d.ts
@@ -2,6 +2,6 @@ import 'express-session';
 
 declare module 'express-session' {
   interface SessionData {
-    dj: { id: number; name: string; email: string };
+    dj: { id: string; name: string; email: string };
   }
 }
